Refactoring operation execution into a service class
