+++ src/Microsoft.Atlas.CommandLine/Execution/IWorkflowEngine.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+using Microsoft.Atlas.CommandLine.Models.Workflow;
+
+namespace Microsoft.Atlas.CommandLine.Execution
+{
+    public interface IWorkflowEngine
+    {
+        Task<object> ExecuteOperations(OperationContext parentContext, IList<WorkflowModel.Operation> operations);
+    }
+}
